+++ cmd/resim/commands/test_suites.go
@@ -42,6 +42,20 @@ var (
 		Run:   getTestSuite,
 	}
 
+	archiveTestSuiteCmd = &cobra.Command{
+		Use:   "archive",
+		Short: "archive - Archive a test suite",
+		Long:  ``,
+		Run:   archiveTestSuite,
+	}
+
+	restoreTestSuiteCmd = &cobra.Command{
+		Use:   "restore",
+		Short: "restore - Restore an archived test suite",
+		Long:  ``,
+		Run:   restoreTestSuite,
+	}
+
 	reviseTestSuiteCmd = &cobra.Command{
 		Use:   "revise",
 		Short: "revise - Revise a test suite, updating the name, metrics, or experiences",
@@ -120,6 +134,19 @@ func init() {
 	getTestSuiteCmd.Flags().Int32(testSuiteRevisionKey, -1, "The specific revision of a test suite to retrieve.")
 	getTestSuiteCmd.Flags().Bool(testSuiteAllRevisionKey, false, "Supply this flag to list all revisions of the test suite.")
 	testSuiteCmd.AddCommand(getTestSuiteCmd)
+
+	// Archive Test Suite
+	archiveTestSuiteCmd.Flags().String(testSuiteProjectKey, "", "The name or ID of the project the test suite is associated with.")
+	archiveTestSuiteCmd.MarkFlagRequired(testSuiteProjectKey)
+	archiveTestSuiteCmd.Flags().String(testSuiteKey, "", "The name or ID of the test suite to archive.")
+	archiveTestSuiteCmd.MarkFlagRequired(testSuiteKey)
+	testSuiteCmd.AddCommand(archiveTestSuiteCmd)
+
+	restoreTestSuiteCmd.Flags().String(testSuiteProjectKey, "", "The name or ID of the project to restore the test suite within")
+	restoreTestSuiteCmd.MarkFlagRequired(testSuiteProjectKey)
+	restoreTestSuiteCmd.Flags().String(testSuiteKey, "", "The name or ID of the test suite to restore")
+	restoreTestSuiteCmd.MarkFlagRequired(testSuiteKey)
+	testSuiteCmd.AddCommand(restoreTestSuiteCmd)
 
 	// Revise Test Suite
 	reviseTestSuiteCmd.Flags().Bool(testSuiteGithubKey, false, "Whether to output format in github action friendly format.")
@@ -226,7 +253,7 @@ func createTestSuite(ccmd *cobra.Command, args []string) {
 	}
 
 	for _, experienceName := range allExperienceNames {
-		experienceID := getExperienceID(Client, projectID, experienceName, true)
+		experienceID := getExperienceID(Client, projectID, experienceName, true, false)
 		allExperienceIDs = append(allExperienceIDs, experienceID)
 	}
 
@@ -290,7 +317,7 @@ func reviseTestSuite(ccmd *cobra.Command, args []string) {
 	}
 
 	// Get the existing test suite name:
-	existingTestSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), nil)
+	existingTestSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), nil, false)
 	if existingTestSuite == nil {
 		log.Fatal("unable to find test suite")
 	}
@@ -337,7 +364,7 @@ func reviseTestSuite(ccmd *cobra.Command, args []string) {
 		allExperienceIDs = append(allExperienceIDs, experienceIDs...)
 		allExperienceNames = append(allExperienceNames, experienceNames...)
 		for _, experienceName := range allExperienceNames {
-			experienceID := getExperienceID(Client, projectID, experienceName, true)
+			experienceID := getExperienceID(Client, projectID, experienceName, true, false)
 			allExperienceIDs = append(allExperienceIDs, experienceID)
 		}
 
@@ -376,7 +403,7 @@ func listTestSuites(ccmd *cobra.Command, args []string) {
 
 	var pageToken *string = nil
 
-	var allTestSuites []api.TestSuite
+	allTestSuites := []api.TestSuite{}
 
 	for {
 		response, err := Client.ListTestSuitesWithResponse(
@@ -392,7 +419,7 @@ func listTestSuites(ccmd *cobra.Command, args []string) {
 
 		pageToken = &response.JSON200.NextPageToken
 		if response.JSON200 == nil || response.JSON200.TestSuites == nil {
-			log.Fatal("no test suites")
+			break
 		}
 		allTestSuites = append(allTestSuites, response.JSON200.TestSuites...)
 		if *pageToken == "" {
@@ -400,10 +427,15 @@ func listTestSuites(ccmd *cobra.Command, args []string) {
 		}
 	}
 
+	if len(allTestSuites) == 0 {
+		fmt.Println("no test suites")
+		return
+	}
+
 	OutputJson(allTestSuites)
 }
 
-func actualGetTestSuite(projectID uuid.UUID, testSuiteKeyRaw string, revision *int32) *api.TestSuite {
+func actualGetTestSuite(projectID uuid.UUID, testSuiteKeyRaw string, revision *int32, expectArchived bool) *api.TestSuite {
 	var testSuite *api.TestSuite
 	if testSuiteKeyRaw == "" {
 		log.Fatal("must specify the test suite name or ID")
@@ -424,6 +456,7 @@ func actualGetTestSuite(projectID uuid.UUID, testSuiteKeyRaw string, revision *i
 			response, err := Client.ListTestSuitesWithResponse(context.Background(), projectID, &api.ListTestSuitesParams{
 				PageToken: pageToken,
 				OrderBy:   Ptr("timestamp"),
+				Archived:  Ptr(expectArchived),
 			})
 			if err != nil {
 				log.Fatal("unable to list test suites:", err)
@@ -444,7 +477,11 @@ func actualGetTestSuite(projectID uuid.UUID, testSuiteKeyRaw string, revision *i
 			if response.JSON200.NextPageToken != "" {
 				pageToken = &response.JSON200.NextPageToken
 			} else {
-				log.Fatal("unable to find test suite: ", testSuiteKeyRaw)
+				if expectArchived {
+					log.Fatal("unable to find archived test suite: ", testSuiteKeyRaw)
+				} else {
+					log.Fatal("unable to find test suite: ", testSuiteKeyRaw)
+				}
 			}
 		}
 	}
@@ -466,7 +503,7 @@ func getTestSuite(ccmd *cobra.Command, args []string) {
 	if viper.IsSet(testSuiteRevisionKey) {
 		revision = Ptr(viper.GetInt32(testSuiteRevisionKey))
 	}
-	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision)
+	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision, false)
 
 	if viper.GetBool(testSuiteAllRevisionKey) {
 		response, err := Client.ListTestSuiteRevisionsWithResponse(context.Background(), projectID, testSuite.TestSuiteID, &api.ListTestSuiteRevisionsParams{
@@ -485,6 +522,32 @@ func getTestSuite(ccmd *cobra.Command, args []string) {
 	}
 }
 
+func archiveTestSuite(ccmd *cobra.Command, args []string) {
+	projectID := getProjectID(Client, viper.GetString(testSuiteProjectKey))
+	var revision *int32
+	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision, false)
+
+	response, err := Client.ArchiveTestSuiteWithResponse(context.Background(), projectID, testSuite.TestSuiteID)
+	if err != nil {
+		log.Fatal("failed to archive test suite:", err)
+	}
+	ValidateResponse(http.StatusNoContent, "failed to archive test suite", response.HTTPResponse, response.Body)
+	fmt.Printf("Archived test suite %s successfully!\n", viper.GetString(testSuiteKey))
+}
+
+func restoreTestSuite(ccmd *cobra.Command, args []string) {
+	projectID := getProjectID(Client, viper.GetString(testSuiteProjectKey))
+	var revision *int32
+	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision, true)
+
+	response, err := Client.RestoreTestSuiteWithResponse(context.Background(), projectID, testSuite.TestSuiteID)
+	if err != nil {
+		log.Fatal("failed to restore test suite:", err)
+	}
+	ValidateResponse(http.StatusNoContent, "failed to restore test suite", response.HTTPResponse, response.Body)
+	fmt.Printf("Restored archived test suite %s successfully!\n", viper.GetString(testSuiteKey))
+}
+
 func runTestSuite(ccmd *cobra.Command, args []string) {
 	testSuiteGithub := viper.GetBool(testSuiteGithubKey)
 	projectID := getProjectID(Client, viper.GetString(testSuiteProjectKey))
@@ -492,7 +555,7 @@ func runTestSuite(ccmd *cobra.Command, args []string) {
 	if viper.IsSet(testSuiteRevisionKey) {
 		revision = Ptr(viper.GetInt32(testSuiteRevisionKey))
 	}
-	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision)
+	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision, false)
 	buildID, err := uuid.Parse(viper.GetString(testSuiteBuildIDKey))
 	if err != nil {
 		log.Fatal("failed to parse build ID: ", err)
@@ -638,7 +701,7 @@ func batchesTestSuite(ccmd *cobra.Command, args []string) {
 		revision = Ptr(viper.GetInt32(testSuiteRevisionKey))
 	}
 
-	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision)
+	testSuite := actualGetTestSuite(projectID, viper.GetString(testSuiteKey), revision, false)
 	testSuiteID := testSuite.TestSuiteID
 
 	batches := []api.Batch{}